+++ src/backend/storage/ipc/Makefile
@@ -11,6 +11,6 @@ override CPPFLAGS := -I$(libpq_srcdir) $(CPPFLAGS)
 
 OBJS = barrier.o dsm_impl.o dsm.o ipc.o ipci.o latch.o pmsignal.o procarray.o \
 	procsignal.o shmem.o shmqueue.o shm_mq.o shm_toc.o signalfuncs.o \
-	sinval.o sinvaladt.o standby.o
+	sinval.o sinvaladt.o standby.o lock_free_list.o lock_free_list_test.o
 
 include $(top_srcdir)/src/backend/common.mk
